+++ src/plugins/localstorage/index.ts
@@ -1,0 +1,3 @@
+export * from './localstorage.module';
+export * from './localstorage.plugin';
+export * from './symbols';